lib: omit unnecessary check in `Task.prototype._mayExecute()`
Omitted the check: `isAsyncTask(actual)` in the `else if`.  It is unnecessary
because the previous `if` checks that the actual is async as has not been
executed, if so it return with... but if it passes then that measn it already
executed and to execute it had to have called it's `_mayExecute()` before
executing, therefore making the second `else if` repetitive by calling
`mayExecute()` for an async task again.